feat: 가격을 int로 변경 #4

+++ src/main/kotlin/me/yevgnenll/stock/entity/Stock.kt
@@ -6,11 +6,11 @@ import java.time.LocalDateTime
 data class Stock(
     var id: Long? = null,
     val timestamp: LocalDate,
-    var low: Double,
-    var high: Double,
-    val close: Double,
-    val open: Double,
-    var volume: Double,
+    var low: Int,
+    var high: Int,
+    val close: Int,
+    val open: Int,
+    var volume: Int,
     val createdAt: LocalDateTime,
     var updatedAt: LocalDateTime,
 ) {
@@ -18,11 +18,11 @@ data class Stock(
         timestamp: LocalDate,
         low: Double, high: Double, close: Double, open: Double, volume: Double) : this(
         timestamp = timestamp,
-        low = low,
-        high = high,
-        close = close,
-        open = open,
-        volume = volume,
+        low = low.toInt(),
+        high = high.toInt(),
+        close = close.toInt(),
+        open = open.toInt(),
+        volume = volume.toInt(),
         createdAt = LocalDateTime.now(),
         updatedAt = LocalDateTime.now(),
     )
